+++ src/init.cpp
@@ -21,7 +21,7 @@
 #include "miner.h"
 #include "net.h"
 #include "options.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "script/standard.h"
 #include "scheduler.h"
 #include "timedata.h"
