+++ src/qt/rpcconsole.cpp
@@ -12,8 +12,8 @@
 
 #include "main.h"
 #include "chainparams.h"
-#include "rpcserver.h"
-#include "rpcclient.h"
+#include "rpc/server.h"
+#include "rpc/client.h"
 #include "util.h"
 
 #include <openssl/crypto.h>
