+++ src/rpc/mining.cpp
@@ -14,7 +14,7 @@
 #include "miner.h"
 #include "net.h"
 #include "pow.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "util.h"
 #include "validationinterface.h"
